menambah attribute di model payment

+++ src/models/cart_detail.js
@@ -33,7 +33,7 @@ module.exports = (sequelize, DataTypes) => {
       },
       id_product: DataTypes.INTEGER,
       quantity: DataTypes.INTEGER,
-      total_price: DataTypes.FLOAT,
+      total_price: DataTypes.INTEGER,
     },
     {
       sequelize,
